+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - removed debug prints
 - added option to ignore friendlies / unattackable units (default active)
 - added Target Name Position setting, defaulting to bottom center
+- added Font Options for Outline and Shadow, defaulting to Outline
 
 ## Version 1.0.0-alpha3